app/requirements.txt
Flask==2.2.3
Werkzeug==2.3.8
gunicorn==20.1.0
redis==4.5.1
psycopg2-binary==2.9.5
